Extract box, center from msg to control
