+++ control/center.go
@@ -1,4 +1,4 @@
-package msg
+package control
 
 import (
 	"context"
